# Walk directories in filesystem source enumeration