<!--
Please create an issue to collect feedback prior to feature additions. Please also reference that issue in any PRs.
If possible try to keep PRs scoped to one feature, and add tests for new features.
-->

### Description:
Previously the unit could be a directory, but now it is individual files. This leverages the SourceManager's concurrent units to speed up scanning of large directories.

```
» find /tmp/files -type f | wc -l
  100000

# before
» ./trufflehog filesystem /tmp/files
2024-01-18T12:48:05-08:00       info-0  trufflehog      finished scanning       {"chunks": 100000, "bytes": 588895, "verified_secrets": 0, "unverified_secrets": 0, "scan_duration": "27.942921208s"}
# 27.94s

# after
» ./trufflehog filesystem /tmp/files
2024-01-18T12:47:04-08:00       info-0  trufflehog      finished scanning       {"chunks": 100000, "bytes": 588895, "verified_secrets": 0, "unverified_secrets": 0, "scan_duration": "14.342492916s"}
# 14.34s
```



### Checklist:
* [ ] Tests passing (`make test-community`)?
* [ ] Lint passing (`make lint` this requires [golangci-lint](https://golangci-lint.run/usage/install/#local-installation))?