@dustin-decker 2024-01-30
Looks like this caused a regression
https://github.com/trufflesecurity/trufflehog/issues/2350